fix buy component

+++ programs/pie/src/instructions/mint_basket_token.rs
@@ -61,8 +61,8 @@ pub fn mint_basket_token(ctx: Context<MintBasketTokenContext>, amount: u64) -> R
             .find(|a| a.mint == token_config.mint)
         {
             let possible_mint = user_asset
-                .amount
-                .checked_div(token_config.ratio as u64)
+            .amount
+                .checked_div(token_config.ratio)
                 .unwrap();
             amount_can_mint = amount_can_mint.min(possible_mint);
         }
@@ -76,12 +76,12 @@ pub fn mint_basket_token(ctx: Context<MintBasketTokenContext>, amount: u64) -> R
             .iter_mut()
             .find(|a| a.mint == token_config.mint)
         {
-            let amount_to_deduct = (token_config.ratio as u64)
+            let amount_to_deduct = token_config.ratio 
                 .checked_mul(amount)
                 .ok_or(PieError::InsufficientBalance)?;
             asset.amount = asset
                 .amount
-                .checked_sub(amount_to_deduct)
+                .checked_sub(amount_to_deduct as u64)
                 .ok_or(PieError::InsufficientBalance)?;
         }
     }
